agg: fix some warnings.

+++ agg/inc/agg_line_aa_basics.h
@@ -44,13 +44,13 @@ namespace agg
     //------------------------------------------------------------------line_mr
     AGG_INLINE int line_mr(int x)
     {
-        return x >> (line_subpixel_shift - line_mr_subpixel_shift);
+        return x >> ((int)line_subpixel_shift - (int)line_mr_subpixel_shift);
     }
 
     //-------------------------------------------------------------------line_hr
     AGG_INLINE int line_hr(int x)
     {
-        return x << (line_subpixel_shift - line_mr_subpixel_shift);
+        return x << ((int)line_subpixel_shift - (int)line_mr_subpixel_shift);
     }
 
     //---------------------------------------------------------------line_dbl_hr
